requirements.txt
httpx==0.27.2
dnspython==2.6.1
python-dotenv==1.0.1
PyYAML==6.0.2
